use 1 file for guardrail table
