+++ ui/litellm-dashboard/src/components/guardrails.tsx
@@ -1,11 +1,5 @@
 import React, { useState, useEffect } from "react";
 import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeaderCell,
-  TableRow,
   Card,
   Text,
   Button,
@@ -13,23 +7,11 @@ import {
 } from "@tremor/react";
 import {
   PlusIcon,
-  TrashIcon,
-  SwitchVerticalIcon,
-  ChevronUpIcon,
-  ChevronDownIcon,
 } from "@heroicons/react/outline";
-import { Tooltip, Modal, message } from "antd";
-import {
-  ColumnDef,
-  flexRender,
-  getCoreRowModel,
-  getSortedRowModel,
-  SortingState,
-  useReactTable,
-} from "@tanstack/react-table";
+import { Modal, message } from "antd";
 import { getGuardrailsList, deleteGuardrailCall } from "./networking";
 import AddGuardrailForm from "./guardrails/add_guardrail_form";
-import { getGuardrailLogoAndName } from "./guardrails/guardrail_info_helpers";
+import GuardrailTable from "./guardrails/guardrail_table";
 
 interface GuardrailsPanelProps {
   accessToken: string | null;
@@ -58,9 +40,6 @@ const GuardrailsPanel: React.FC<GuardrailsPanelProps> = ({ accessToken }) => {
   const [isLoading, setIsLoading] = useState(false);
   const [isDeleting, setIsDeleting] = useState(false);
   const [guardrailToDelete, setGuardrailToDelete] = useState<{id: string, name: string} | null>(null);
-  const [sorting, setSorting] = useState<SortingState>([
-    { id: "created_at", desc: true }
-  ]);
 
   const fetchGuardrails = async () => {
     if (!accessToken) {
@@ -120,140 +99,6 @@ const GuardrailsPanel: React.FC<GuardrailsPanelProps> = ({ accessToken }) => {
     setGuardrailToDelete(null);
   };
 
-  // Format date helper function
-  const formatDate = (dateString?: string) => {
-    if (!dateString) return "-";
-    const date = new Date(dateString);
-    return date.toLocaleString();
-  };
-
-  const columns: ColumnDef<GuardrailItem>[] = [
-    {
-      header: "Name",
-      accessorKey: "guardrail_name",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        return (
-          <Tooltip title={guardrail.guardrail_name}>
-            <span className="text-xs font-medium">
-              {guardrail.guardrail_name || "-"}
-            </span>
-          </Tooltip>
-        );
-      },
-    },
-    {
-      header: "Provider",
-      accessorKey: "litellm_params.guardrail",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        const { logo, displayName } = getGuardrailLogoAndName(guardrail.litellm_params.guardrail);
-        return (
-          <div className="flex items-center space-x-2">
-            {logo && (
-              <img 
-                src={logo} 
-                alt={`${displayName} logo`} 
-                className="w-4 h-4"
-                onError={(e) => {
-                  // Hide broken image
-                  (e.target as HTMLImageElement).style.display = 'none';
-                }}
-              />
-            )}
-            <span className="text-xs">{displayName}</span>
-          </div>
-        );
-      },
-    },
-    {
-      header: "Mode",
-      accessorKey: "litellm_params.mode",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        return (
-          <span className="text-xs">
-            {guardrail.litellm_params.mode}
-          </span>
-        );
-      },
-    },
-    {
-      header: "Status",
-      accessorKey: "litellm_params.default_on",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        return (
-          <div className={`inline-flex rounded-full px-2 py-1 text-xs font-medium
-              ${guardrail.litellm_params.default_on 
-              ? 'bg-green-100 text-green-800'  // Always On styling
-              : 'bg-gray-100 text-gray-800'    // Per Request styling
-              }`}>
-              {guardrail.litellm_params.default_on ? 'Always On' : 'Per Request'}
-          </div>
-        );
-      },
-    },
-    {
-      header: "Created At",
-      accessorKey: "created_at",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        return (
-          <Tooltip title={guardrail.created_at}>
-            <span className="text-xs">
-              {formatDate(guardrail.created_at)}
-            </span>
-          </Tooltip>
-        );
-      },
-    },
-    {
-      header: "Updated At",
-      accessorKey: "updated_at",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        return (
-          <Tooltip title={guardrail.updated_at}>
-            <span className="text-xs">
-              {formatDate(guardrail.updated_at)}
-            </span>
-          </Tooltip>
-        );
-      },
-    },
-    {
-      id: "actions",
-      header: "",
-      cell: ({ row }) => {
-        const guardrail = row.original;
-        return (
-          <div className="flex space-x-2">
-            <Icon
-              icon={TrashIcon}
-              size="sm"
-              onClick={() => guardrail.guardrail_id && handleDeleteClick(guardrail.guardrail_id, guardrail.guardrail_name || 'Unnamed Guardrail')}
-              className="cursor-pointer hover:text-red-500"
-              tooltip="Delete guardrail"
-            />
-          </div>
-        );
-      },
-    },
-  ];
-
-  const table = useReactTable({
-    data: guardrailsList,
-    columns,
-    state: {
-      sorting,
-    },
-    onSortingChange: setSorting,
-    getCoreRowModel: getCoreRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    enableSorting: true,
-  });
-
   return (
     <div className="w-full mx-auto flex-auto overflow-y-auto m-8 p-2">
       <div className="flex justify-between items-center mb-4">
@@ -265,88 +110,12 @@ const GuardrailsPanel: React.FC<GuardrailsPanelProps> = ({ accessToken }) => {
           Add Guardrail
         </Button>
       </div>
-        <div className="rounded-lg custom-border relative">
-          <div className="overflow-x-auto">
-            <Table className="[&_td]:py-0.5 [&_th]:py-1">
-              <TableHead>
-                {table.getHeaderGroups().map((headerGroup) => (
-                  <TableRow key={headerGroup.id}>
-                    {headerGroup.headers.map((header) => (
-                      <TableHeaderCell
-                        key={header.id}
-                        className={`py-1 h-8 ${
-                          header.id === 'actions'
-                            ? 'sticky right-0 bg-white shadow-[-4px_0_8px_-6px_rgba(0,0,0,0.1)]'
-                            : ''
-                        }`}
-                        onClick={header.column.getToggleSortingHandler()}
-                      >
-                        <div className="flex items-center justify-between gap-2">
-                          <div className="flex items-center">
-                            {header.isPlaceholder ? null : (
-                              flexRender(
-                                header.column.columnDef.header,
-                                header.getContext()
-                              )
-                            )}
-                          </div>
-                          {header.id !== 'actions' && (
-                            <div className="w-4">
-                              {header.column.getIsSorted() ? (
-                                {
-                                  asc: <ChevronUpIcon className="h-4 w-4 text-blue-500" />,
-                                  desc: <ChevronDownIcon className="h-4 w-4 text-blue-500" />
-                                }[header.column.getIsSorted() as string]
-                              ) : (
-                                <SwitchVerticalIcon className="h-4 w-4 text-gray-400" />
-                              )}
-                            </div>
-                          )}
-                        </div>
-                      </TableHeaderCell>
-                    ))}
-                  </TableRow>
-                ))}
-              </TableHead>
-              <TableBody>
-                {isLoading ? (
-                  <TableRow>
-                    <TableCell colSpan={columns.length} className="h-8 text-center">
-                      <div className="text-center text-gray-500">
-                        <p>Loading...</p>
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                ) : table.getRowModel().rows.length > 0 ? (
-                  table.getRowModel().rows.map((row) => (
-                    <TableRow key={row.id} className="h-8">
-                      {row.getVisibleCells().map((cell) => (
-                        <TableCell
-                          key={cell.id}
-                          className={`py-0.5 max-h-8 overflow-hidden text-ellipsis whitespace-nowrap ${
-                            cell.column.id === 'actions'
-                              ? 'sticky right-0 bg-white shadow-[-4px_0_8px_-6px_rgba(0,0,0,0.1)]'
-                              : ''
-                          }`}
-                        >
-                          {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                        </TableCell>
-                      ))}
-                    </TableRow>
-                  ))
-                ) : (
-                  <TableRow>
-                    <TableCell colSpan={columns.length} className="h-8 text-center">
-                      <div className="text-center text-gray-500">
-                        <p>No guardrails found</p>
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                )}
-              </TableBody>
-            </Table>
-          </div>
-        </div>
+      
+      <GuardrailTable 
+        guardrailsList={guardrailsList}
+        isLoading={isLoading}
+        onDeleteClick={handleDeleteClick}
+      />
 
       <AddGuardrailForm 
         visible={isAddModalVisible}
